backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.1.7
python-multipart==0.0.9
pydantic-settings==2.2.1
uuid7==0.1.0
email-validator==2.1.1
greenlet==3.0.3
google-auth==2.48.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
google-api-python-client==2.192.0
celery==5.6.2
redis==7.3.0
openai==1.14.3